+++ src/test/java/io/github/astrapi69/collection/map/MapExtensionsTest.java
@@ -33,9 +33,9 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Properties;
 import java.util.Set;
-import java.util.Map.Entry;
 
 import org.meanbean.test.BeanTester;
 import org.testng.Assert;